+++ src/pybamm/expression_tree/interpolant.py
@@ -107,6 +107,9 @@ class Interpolant(pybamm.Function):
                     "len(x1) should equal y=shape[0], "
                     f"but x1.shape={x1.shape} and y.shape={y.shape}"
                 )
+            # if 1d then x should be monotonically increasing
+            if np.any(x1[:-1] > x1[1:]):
+                raise ValueError("x should be monotonically increasing")
         # children should be a list not a symbol or a number
         if isinstance(children, pybamm.Symbol | numbers.Number):
             children = [children]